src/requirements.txt
Gooey
Pillow
